requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-docx
PyPDF2
transformers
torch
openai
python-dotenv